Moving functionality out of brewtils init
