+++ test/config_test.py
@@ -1,18 +1,15 @@
 # -*- coding: utf-8 -*-
-
 import copy
 import os
 import warnings
 
 import pytest
 
-import brewtils
-import brewtils.rest
+from brewtils.config import load_config, get_argument_parser, get_connection_info
 from brewtils.errors import ValidationError
-from brewtils.rest.easy_client import EasyClient
 
 
-class TestBrewtils(object):
+class TestGetConnectionInfo(object):
     @pytest.fixture
     def params(self):
         return {
@@ -37,50 +34,10 @@ class TestBrewtils(object):
     def teardown_method(self):
         os.environ = self.safe_copy
 
-    def test_load_config_cli(self):
-        cli_args = ["--bg-host", "the_host"]
+    def test_kwargs(self, params):
+        assert params == get_connection_info(**params)
 
-        config = brewtils.load_config(cli_args)
-        assert config.bg_host == "the_host"
-
-    def test_load_config_cli_custom_argument_parser_vars(self):
-        parser = brewtils.get_argument_parser()
-        parser.add_argument("some_parameter")
-
-        cli_args = ["param", "--bg-host", "the_host"]
-        parsed_args = parser.parse_args(cli_args)
-
-        config = brewtils.load_config(**vars(parsed_args))
-        assert config.bg_host == "the_host"
-        assert parsed_args.some_parameter == "param"
-        assert "some_parameter" not in config
-
-    def test_load_config_cli_custom_argument_parser_cli(self):
-        parser = brewtils.get_argument_parser()
-        parser.add_argument("some_parameter")
-
-        cli_args = ["param", "--bg-host", "the_host"]
-        parsed_args = parser.parse_args(cli_args)
-
-        config = brewtils.load_config(cli_args=cli_args, argument_parser=parser)
-        assert config.bg_host == "the_host"
-        assert parsed_args.some_parameter == "param"
-        assert "some_parameter" not in config
-
-    def test_load_config_environment(self):
-        os.environ["BG_HOST"] = "the_host"
-
-        config = brewtils.load_config([])
-        assert config.bg_host == "the_host"
-
-    def test_get_easy_client(self):
-        client = brewtils.get_easy_client(bg_host="bg_host")
-        assert isinstance(client, EasyClient) is True
-
-    def test_get_connection_info_kwargs(self, params):
-        assert params == brewtils.get_connection_info(**params)
-
-    def test_get_connection_info_env(self, params):
+    def test_env(self, params):
         os.environ["BG_HOST"] = "bg_host"
         os.environ["BG_PORT"] = "1234"
         os.environ["BG_SSL_ENABLED"] = "False"
@@ -89,33 +46,33 @@ class TestBrewtils(object):
         os.environ["BG_URL_PREFIX"] = "/beer/"
         os.environ["BG_CA_VERIFY"] = "True"
 
-        assert params == brewtils.get_connection_info()
+        assert params == get_connection_info()
 
-    def test_get_connection_info_deprecated_kwarg_host(self, params):
+    def test_deprecated_kwarg_host(self, params):
         deprecated_params = copy.copy(params)
         deprecated_params["host"] = deprecated_params.pop("bg_host")
 
         with warnings.catch_warnings(record=True) as w:
             warnings.simplefilter("always")
 
-            assert params == brewtils.get_connection_info(**deprecated_params)
+            assert params == get_connection_info(**deprecated_params)
 
             assert issubclass(w[0].category, DeprecationWarning)
             assert "host" in str(w[0].message)
 
-    def test_get_connection_info_deprecated_kwarg_port(self, params):
+    def test_deprecated_kwarg_port(self, params):
         deprecated_params = copy.copy(params)
         deprecated_params["port"] = deprecated_params.pop("bg_port")
 
         with warnings.catch_warnings(record=True) as w:
             warnings.simplefilter("always")
 
-            assert params == brewtils.get_connection_info(**deprecated_params)
+            assert params == get_connection_info(**deprecated_params)
 
             assert issubclass(w[0].category, DeprecationWarning)
             assert "port" in str(w[0].message)
 
-    def test_get_connection_info_deprecated_env(self, params):
+    def test_deprecated_env(self, params):
         deprecated_params = copy.copy(params)
         deprecated_params["bg_host"] = None
         deprecated_params["bg_port"] = None
@@ -127,15 +84,59 @@ class TestBrewtils(object):
         os.environ["BG_WEB_HOST"] = "bg_host"
         os.environ["BG_WEB_PORT"] = "1234"
 
-        assert params == brewtils.get_connection_info(**deprecated_params)
+        assert params == get_connection_info(**deprecated_params)
 
-    def test_get_connection_info_no_host(self):
+    def test_no_host(self):
         with pytest.raises(ValidationError):
-            brewtils.get_connection_info()
+            get_connection_info()
 
     def test_normalize_url_prefix(self, params):
         os.environ["BG_WEB_HOST"] = "bg_host"
         os.environ["BG_URL_PREFIX"] = "/beer"
 
-        generated_params = brewtils.get_connection_info()
+        generated_params = get_connection_info()
         assert generated_params["url_prefix"] == params["url_prefix"]
+
+
+class TestLoadConfig(object):
+    def setup_method(self):
+        self.safe_copy = os.environ.copy()
+
+    def teardown_method(self):
+        os.environ = self.safe_copy
+
+    def test_cli(self):
+        cli_args = ["--bg-host", "the_host"]
+
+        config = load_config(cli_args)
+        assert config.bg_host == "the_host"
+
+    def test_cli_custom_argument_parser_vars(self):
+        parser = get_argument_parser()
+        parser.add_argument("some_parameter")
+
+        cli_args = ["param", "--bg-host", "the_host"]
+        parsed_args = parser.parse_args(cli_args)
+
+        config = load_config(**vars(parsed_args))
+        assert config.bg_host == "the_host"
+        assert parsed_args.some_parameter == "param"
+        assert "some_parameter" not in config
+
+    def test_cli_custom_argument_parser_cli(self):
+        parser = get_argument_parser()
+        parser.add_argument("some_parameter")
+
+        cli_args = ["param", "--bg-host", "the_host"]
+        parsed_args = parser.parse_args(cli_args)
+
+        config = load_config(cli_args=cli_args, argument_parser=parser)
+        assert config.bg_host == "the_host"
+        assert parsed_args.some_parameter == "param"
+        assert "some_parameter" not in config
+
+    def test_environment(self):
+        os.environ["BG_HOST"] = "the_host"
+
+        config = load_config([])
+        assert config.bg_host == "the_host"
